feat: add route protection middleware
