+++ auth.config.ts
@@ -1,0 +1,6 @@
+import Google from "next-auth/providers/google";
+import type { NextAuthConfig } from "next-auth";
+
+export default {
+  providers: [Google],
+} satisfies NextAuthConfig;
