+++ package.json
@@ -17,7 +17,7 @@
     "@liveblocks/react-tiptap": "^3.4.1",
     "@liveblocks/react-ui": "^3.4.1",
     "@liveblocks/yjs": "^3.4.1",
-    "@prisma/client": "^6.4.1",
+    "@prisma/client": "^6.17.1",
     "@radix-ui/react-alert-dialog": "^1.1.6",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-checkbox": "^1.3.3",
@@ -57,7 +57,7 @@
     "lucide-react": "^0.475.0",
     "mammoth": "^1.11.0",
     "next": "^15.5.2",
-    "next-auth": "^5.0.0-beta.25",
+    "next-auth": "^5.0.0-beta.29",
     "next-themes": "^0.4.6",
     "nodemailer": "^6.10.0",
     "openai": "^4.85.3",
