+++ prisma/schema.prisma
@@ -18,7 +18,6 @@ model User {
   phoneNumber   String?
   accounts      Account[]
   sessions      Session[]
-  Exam Exam[]
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 
@@ -73,28 +72,6 @@ model VerificationToken {
   @@unique([identifier, token])
 }
 
-
-model Exam {
-  id String  @id @default(auto()) @map("_id") @db.ObjectId
-  userId String @db.ObjectId
-  title String
-  questions Question[]
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-
-  user User @relation(fields: [userId], references: [id], onDelete:Cascade)
-}
-
-
-
-model Question {
-  id String  @id @default(auto()) @map("_id") @db.ObjectId
-  examId String @db.ObjectId
-  content String
-  createdAt DateTime @default(now())
-
-  exam Exam @relation(fields: [examId], references: [id], onDelete:Cascade)
-}
 
 model News {
   id String @id @default(auto()) @map("_id") @db.ObjectId
